Joing columns

+++ src/main/java/com/example/monara_backend/repository/PdfBillRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PdfBillRepo extends JpaRepository<PdfBillSave,Integer> {
+public interface PdfBillRepo extends JpaRepository<PdfBillSave,Long> {
 
 }
